+++ client/src/components/MainContent.tsx
@@ -1,21 +1,23 @@
 import { useState } from 'react';
+// @ts-ignore: No Implicit Any - This is a js file
+import useLocalStorage from '../hooks/useLocalStorage';
 import { ShortenedLink } from '../models/ShortenedLink';
-import ShortenedLinksList from './ShortenLink/ShortenedLinksList';
-import ShortenLinkForm2 from './ShortenLink/ShortenLinkForm2';
+import ShortenedLinksList from './ShortenedLinksList';
+import ShortenLinkForm from './ShortenLinkForm';
 import Title from './ui/Title';
 
 const MainContent: React.FC = () => {
   const [shortenedLinks, setShortenedLinks] = useState<ShortenedLink[]>([]);
 
   const addShortenedLink = (shortenedLink: ShortenedLink) => {
-    setShortenedLinks((prev) => [shortenedLink, ...prev]);
+    setShortenedLinks((prev: ShortenedLink[]) => [shortenedLink, ...prev]);
   };
 
   return (
     <>
-      <div className='mt-[20%]'>
+      <div className='mt-[15%]'>
         <Title>Purlto</Title>
-        <ShortenLinkForm2 addShortenedLink={addShortenedLink} />
+        <ShortenLinkForm addShortenedLink={addShortenedLink} />
         <ShortenedLinksList shortenedLinks={shortenedLinks} />
       </div>
     </>
