--- client/src/components/icons/index.tsx
@@ -1,7 +1,0 @@
-import CopyIcon from './CopyIcon';
-
-const Icon = {
-  CopyIcon: CopyIcon,
-};
-
-export default Icon;
