+++ client/src/hooks/useLocalStorage.jsx
@@ -1,23 +1,27 @@
-// source: https://usehooks.com/useLocalStorage/
+import { useState } from 'react';
 
-import { useEffect, useState } from 'react';
-
-export default function useLocalStorage<T>(key: string, initialValue: T) {
+// Hook
+export default function useLocalStorage(key, initialValue) {
   // State to store our value
-  const [storedValue, setStoredValue] = useState<T>(initialValue);
-
-  useEffect(() => {
-    // Get from local storage by key
-    const item = window.localStorage.getItem(key);
-
-    if (item) {
-      setStoredValue(JSON.parse(item));
+  // Pass initial state function to useState so logic is only executed once
+  const [storedValue, setStoredValue] = useState(() => {
+    if (typeof window === 'undefined') {
+      return initialValue;
     }
-  }, [key]);
-
+    try {
+      // Get from local storage by key
+      const item = window.localStorage.getItem(key);
+      // Parse stored json or if none return initialValue
+      return item ? JSON.parse(item) : initialValue;
+    } catch (error) {
+      // If error also return initialValue
+      console.log(error);
+      return initialValue;
+    }
+  });
   // Return a wrapped version of useState's setter function that ...
   // ... persists the new value to localStorage.
-  const setValue = (value: T) => {
+  const setValue = (value) => {
     try {
       // Allow value to be a function so we have same API as useState
       const valueToStore =
