+++ client/src/main.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import * as ReactDOM from 'react-dom/client';
 import {
   createBrowserRouter,
@@ -6,21 +5,16 @@ import {
   RouterProvider,
 } from 'react-router-dom';
 import './assets/css/index.css';
-import App from './App';
-import PageNotFound from './error-page';
+import PageNotFound from './pages/PageNotFound';
 import Feedback from './pages/Feedback';
 import Redirect, { loader as redirectLoader } from './pages/Redirect';
-import App2 from './App2';
+import App from './App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 
 const router = createBrowserRouter([
-  {
-    path: '/2',
-    element: <App2 />,
-  },
   {
     path: '/',
     element: <App />,
