+++ client/src/pages/PageNotFound.tsx
@@ -1,8 +1,7 @@
-import Footer from '../components/layout/Footer';
-import Header from '../components/layout/Header';
 import BackgroundBubbles from '../components/ui/BackgroundBubbles';
-
-const Feedback = () => {
+import Header from '../components/layout/Header';
+import Footer from '../components/layout/Footer';
+const PageNotFound = () => {
   return (
     <>
       <div className='relative overflow-hidden bg-gradient-to-bl via-indigo-blue to-steel-blue from-indigo-purple '>
@@ -10,9 +9,7 @@ const Feedback = () => {
         <main className='container mx-auto min-h-screen flex flex-col px-2'>
           <Header />
           <section className='flex-1 flex flex-col items-center justify-center'>
-            <h1 className='text-7xl text-white font-bold'>
-              Under Construction...
-            </h1>
+            <h1 className='text-7xl text-white font-bold'>404</h1>
           </section>
           <Footer />
         </main>
@@ -21,4 +18,4 @@ const Feedback = () => {
   );
 };
 
-export default Feedback;
+export default PageNotFound;
